[FEATURE/CLEANUP] converted card formatting to utility
